requirements.txt
annotated-types==0.5.0
anyio==3.7.1
beautifulsoup4==4.12.2
boto3==1.33.6
botocore==1.33.6
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.7
contourpy==1.1.1
cycler==0.11.0
dill==0.3.7
dnspython==2.4.2
exceptiongroup==1.1.3
fastapi==0.103.1
filelock==3.9.0
fonttools==4.42.1
gdown==4.7.1
h11==0.14.0
idna==3.4
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.3.2
kiwisolver==1.4.5
lightgbm==4.1.0
MarkupSafe==2.1.2
matplotlib==3.8.0
mpmath==1.3.0
networkx==3.0
numpy==1.26.0
opencv-python==4.8.1.78
packaging==23.1
pandas==2.1.1
Pillow==9.3.0
psutil==5.9.5
py-cpuinfo==9.0.0
pydantic==2.4.2
pydantic_core==2.10.1
pymongo==4.5.0
pyparsing==3.1.1
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.28.1
s3transfer==0.8.2
scikit-learn==1.3.0
scipy==1.11.3
seaborn==0.12.2
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
starlette==0.27.0
sympy==1.12
threadpoolctl==3.2.0
torch==2.0.1+cpu
torchaudio==2.0.2+cpu
torchvision==0.15.2+cpu
tqdm==4.66.1
typing_extensions==4.8.0
tzdata==2023.3
ultralytics==8.0.188
urllib3==1.26.13
uvicorn==0.23.2
